fix(jpa): use BigDecimal for hazard coordinates to satisfy Hibernate size rules

+++ src/main/java/com/example/villagefirerisk/entity/Hazard.java
@@ -2,6 +2,7 @@ package com.example.villagefirerisk.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -28,10 +29,10 @@ public class Hazard extends BaseEntity {
     private String address;
 
     @Column(nullable = false, precision = 10, scale = 7)
-    private Double latitude;
+    private BigDecimal latitude;
 
     @Column(nullable = false, precision = 10, scale = 7)
-    private Double longitude;
+    private BigDecimal longitude;
 
     @Column(name = "image_url", length = 500)
     private String imageUrl;
@@ -77,10 +78,10 @@ public class Hazard extends BaseEntity {
     public void setAreaCode(String areaCode) { this.areaCode = areaCode; }
     public String getAddress() { return address; }
     public void setAddress(String address) { this.address = address; }
-    public Double getLatitude() { return latitude; }
-    public void setLatitude(Double latitude) { this.latitude = latitude; }
-    public Double getLongitude() { return longitude; }
-    public void setLongitude(Double longitude) { this.longitude = longitude; }
+    public BigDecimal getLatitude() { return latitude; }
+    public void setLatitude(BigDecimal latitude) { this.latitude = latitude; }
+    public BigDecimal getLongitude() { return longitude; }
+    public void setLongitude(BigDecimal longitude) { this.longitude = longitude; }
     public String getImageUrl() { return imageUrl; }
     public void setImageUrl(String imageUrl) { this.imageUrl = imageUrl; }
     public HazardStatus getStatus() { return status; }
